rename validationUtils toUpperCase
