+++ frontend/src/components/messstelle/MessquerschnittForm.vue
@@ -112,7 +112,7 @@ import { onMounted, ref, watch } from "vue";
 import LhmTextField from "@/components/common/LhmTextField.vue";
 import { useEventbus } from "@/store/Eventbus";
 import { himmelsRichtungenTextLong } from "@/types/enum/Himmelsrichtungen";
-import { useValidationUtils } from "@/util/validationUtils";
+import { useValidationUtils } from "@/util/ValidationUtils";
 
 const validationUtils = useValidationUtils();
